Change main file name to MeowBot
